Move legend to widgets/legend
