+++ src/entities/legend_item/index.ts
