+++ src/widgets/legend/index.tsx
@@ -1,15 +1,15 @@
 import { memo } from 'react';
-import { LegendItem as Item } from 'shared/types';
+import { LegendItem } from 'entities/legend_item';
 
-type LegendItemProps = Item & {
+type LegendItemProps = LegendItem & {
   id: string;
 };
 
 type Props = {
-  items: Item[];
+  items: LegendItem[];
 };
 
-const LegendItem = ({ id, color, label }: LegendItemProps) => {
+const Item = ({ id, color, label }: LegendItemProps) => {
   const className = color + 'w-[1rem] h-[1rem] mask mask-circle mr-1';
   return (
     <div id={id} className='mr-auto flex flex-row items-center'>
@@ -23,7 +23,7 @@ const Component = ({ items }: Props) => {
   return (
     <div id='legend' className='flex flex-col'>
       {items.map((item, index) => (
-        <LegendItem
+        <Item
           id={`legend_item[${index}]`}
           key={index}
           color={item.color}
